Moved auth function to login activity
