+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,8 @@
             </intent-filter>
         </activity>
 
+        <activity android:name=".ui.main.LauncherActivity"/>
+        <activity android:name=".ui.main.LoginActivity"/>
 
     </application>
