+++ app/src/main/java/com/example/taskplanner/ui/main/MainActivity.kt
@@ -15,36 +15,14 @@ import kotlinx.coroutines.IO
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
-
-    @Inject
-    lateinit var authService: AuthService
-
-    @Inject
-    lateinit var storage: Storage
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        auth()
+
 
     }
 
-    private fun auth() {
 
-        GlobalScope.launch(Dispatchers.IO) {
-            val response = authService.auth(LoginDto("davidcab11@gmail.com","passw0rd"))
-            if(response.isSuccessful){
-                val tokenDto = response.body()!!
-                Log.d("DEBUG", "tokenDto: $tokenDto")
-                storage.saveToken(tokenDto.token)
-            }else{
-                response.errorBody()
-            }
-
-        }
-
-     }
 }
